requirements.txt
django==3.1.1
django-environ==0.4.5
psycopg2==2.8.6
Pillow==7.2.0
djangorestframework==3.11.1
whitenoise==5.2.0
django-widget-tweaks==1.4.8
gunicorn==20.0.4
dj-database-url==0.5.0
django-filter==2.4.0
django-cors-headers==3.5.0